# perf: Misc perf improvement to ResourceDictionary